requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anthropic==0.46.0
    # via marker-pdf
anyio==4.10.0
    # via
    #   anthropic
    #   google-genai
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
beautifulsoup4==4.13.5
    # via
    #   -r requirements.in
    #   markdownify
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   marker-pdf
    #   pdftext
    #   surya-ocr
    #   uvicorn
distlib==0.4.0
    # via virtualenv
distro==1.9.0
    # via
    #   anthropic
    #   openai
einops==0.8.1
    # via surya-ocr
fastapi==0.116.1
    # via -r requirements.in
filelock==3.19.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   virtualenv
filetype==1.2.0
    # via
    #   marker-pdf
    #   surya-ocr
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
ftfy==6.3.1
    # via marker-pdf
google-auth==2.40.3
    # via google-genai
google-genai==1.36.0
    # via marker-pdf
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.10
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   anthropic
    #   google-genai
    #   ollama
    #   openai
huggingface-hub==0.34.4
    # via
    #   tokenizers
    #   transformers
identify==2.6.14
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via torch
jiter==0.10.0
    # via
    #   anthropic
    #   openai
joblib==1.5.2
    # via scikit-learn
markdown2==2.5.4
    # via marker-pdf
markdownify==1.2.0
    # via marker-pdf
marker-pdf[ocr]==1.9.3
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
nodeenv==1.9.1
    # via pre-commit
numpy==2.3.3
    # via
    #   -r requirements.in
    #   opencv-python-headless
    #   scikit-learn
    #   scipy
    #   transformers
ollama==0.5.3
    # via -r requirements.in
openai==1.107.2
    # via
    #   -r requirements.in
    #   marker-pdf
opencv-python-headless==4.11.0.86
    # via surya-ocr
packaging==25.0
    # via
    #   huggingface-hub
    #   pytest
    #   transformers
pdftext==0.6.3
    # via marker-pdf
pillow==10.4.0
    # via
    #   -r requirements.in
    #   marker-pdf
    #   surya-ocr
platformdirs==4.4.0
    # via
    #   surya-ocr
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.3.0
    # via
    #   marker-pdf
    #   surya-ocr
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   anthropic
    #   fastapi
    #   google-genai
    #   marker-pdf
    #   ollama
    #   openai
    #   pdftext
    #   pydantic-settings
    #   surya-ocr
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via
    #   marker-pdf
    #   pdftext
    #   surya-ocr
pygments==2.19.2
    # via pytest
pypdfium2==4.30.0
    # via
    #   pdftext
    #   surya-ocr
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   marker-pdf
    #   pydantic-settings
    #   surya-ocr
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pre-commit
    #   transformers
    #   uvicorn
rapidfuzz==3.14.1
    # via marker-pdf
regex==2024.11.6
    # via
    #   marker-pdf
    #   transformers
requests==2.32.5
    # via
    #   google-genai
    #   huggingface-hub
    #   transformers
rsa==4.9.1
    # via google-auth
safetensors==0.6.2
    # via transformers
scikit-learn==1.7.2
    # via marker-pdf
scipy==1.16.2
    # via scikit-learn
six==1.17.0
    # via markdownify
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
starlette==0.47.3
    # via fastapi
surya-ocr==0.16.7
    # via marker-pdf
sympy==1.14.0
    # via
    #   -r requirements.in
    #   torch
tenacity==9.1.2
    # via
    #   -r requirements.in
    #   google-genai
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.0
    # via transformers
torch==2.8.0
    # via
    #   marker-pdf
    #   surya-ocr
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   marker-pdf
    #   openai
    #   transformers
transformers==4.56.1
    # via
    #   marker-pdf
    #   surya-ocr
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   google-genai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   starlette
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
virtualenv==20.34.0
    # via pre-commit
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via ftfy
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
